// src/lib/generated/requisition/models/periodType.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Kaizen Admin Approval System
 * Kaizen Admin and Approval Management System using LangGraph
 * OpenAPI spec version: 1.0.0
 */

/**
 * Budget period types
 */
export type PeriodType = typeof PeriodType[keyof typeof PeriodType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PeriodType = {
  monthly: 'monthly',
  quarterly: 'quarterly',
  semi_annual: 'semi_annual',
  annual: 'annual',
  custom: 'custom',
} as const;